feat: add JsonConverterFactory
